requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord.py==2.4.0
dnspython==2.6.1
facebook-sdk==3.1.0
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pymongo==4.8.0
pyTelegramBotAPI==4.23.0
python-telegram-bot==21.4
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
sniffio==1.3.1
soupsieve==2.6
telebot==0.0.5
tokenizers==0.19.1
tqdm==4.66.5
transformers==4.43.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.11
